+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Access to metadata for the api metadata `implementation`'s members
+
+### Changed
+- Api metadata's `members` map deprecated and will be removed in the next major version. All its data is now in `implementation`
+
 ## 2.3.0 - 2024-03-13
 Names and paths for the `apiS` metadata. Documentation for `apiS` part of the library